requirements.txt
Django==5.0.3
django-versatileimagefield==0.2.1
djangorestframework==3.15.0
django-cors-headers==4.3.1
asgiref
python-decouple
pytz
psycopg2-binary
sqlparse
Pillow
tzdata
mysqlclient
drf_yasg